Improved reading API and preliminary writing support (WIP)
Sorry for this clusterfuck of a commit, it's another one of those that I took a break from for awhile and I forgot all the changes I made. Code seems good and everything checks out in the tests, so it should be fine.

+++ Marathon/Formats/Ninja/Chunks/TextureListChunk.cs
@@ -54,7 +54,7 @@ namespace Marathon.Formats.Ninja.Chunks
             var header = new ChunkHeader(in_writer, GetChunkID(), 0);
 
             var textureFilePos = (int)(in_writer.Position - InfoChunk.Size);
-            var textureFileNameOffsets = new List<uint>();
+            var textureFileNameOffsets = new List<long>();
 
             foreach (var texture in Textures)
             {
